Add files via upload
basically just add this as csv on the website, and you'll recieve analytics for the same any other csv file not in the format (acc to model specs-->although if there are similar cols; analytics will be visible,  just have a look thru this dataset and curate your dataset based on it). Testing part of the dataset is this csv file, the model is already trained with a csv file that is mentioned in the repo.